==> 01. S1a Sensor and Activities Information.ipynb <==
import pandas as pd
import numpy as np
import json

---
<div style="text-align: justify">

## The Activities Data
### Importing the Activities Data
The file `S1Activities.csv` contains a tabulated summary of Heading, Category, Subcategory and a corresponding code.

dsActivities = pd.read_csv('S1Activities.csv', index_col = None)    # Importing the dataset

dsActivities.head(n=10)                                             # Checking the dataset

len(dsActivities)

At this time, the activities data will not be subject to preprocessing.

## The Sensor Data
### Importing Sensor Data
The file `S1sensors.csv` contains a tabulated values for sensor ID (numeric), room and sensor type.

dsS1Sensors = pd.read_csv('S1sensors.csv',                    # Importing the dataset
                          index_col = None, header = None)    

dsS1Sensors.head(n=5)                                         # Checking the dataset

len(dsS1Sensors)                                      # Checking the length of dsS1Sensors

### Preprocessing the Sensor Data
Column [1] & Column [2] of the sensor data will be concatenated, whitespace will be removed, all text will be cast to lowercase and a final whitespace strip will be performed. 

temp1 = dsS1Sensors[1] + '_' + dsS1Sensors[2]   # Creating a new vector of concatenated column 1 & 2
temp2 = temp1.str.replace(" ", "")              # Removing whitespace between strings
temp3 = temp2.str.lower()                       # Changing all text to lowercase
temp4 = temp3.str.strip()                       # Striping any remaining whitespace
dsS1Sensors[3] = temp4                          # Adding the precessed vector back into the dataset
dsS1Sensors.rename(columns={0:'subActNum', 
                            1:'room',
                            2:'activity',
                            3:'concat',
                            4:'energyReq'}, 
                   inplace=True)

dsS1Sensors.head(n=6)                           # Checking the resultant dataset

<p style='text-align: justify;'> 
Each row needs to be inspected manually to determine if the activity requires electricity. This can only be done initially by manual inspection.
</p>

<p style='text-align: justify;'> 
</p>

def showUniqueValues(list1):                 # Function to get unique values 
    unique_list = []                         # Intilize a null list 
    
    for x in list1:                          # Traverse for all elements 
        if x not in unique_list:             # Check if exists in unique_list or not 
            unique_list.append(x)            # Append to unique_list
    
    for x in unique_list:                    # Print list 
        print(x)

showUniqueValues(dsS1Sensors.iloc[:,3])   # Pass the concacenated sensor list to unique

a = dsS1Sensors.iloc[:,3]
specialChar = '/'
reqElectricity = 'ligh|burn|mach|toas|freez|dvd|lamp|washer|dry|exh|disp|frig|oven|hot|micro'
b = a.str.contains(reqElectricity, regex = True)
c = a.str.contains(specialChar, regex = True)
dsS1Sensors['reqElectricity'] = b
dsS1Sensors['specialChar'] = c

forwardSlashFilter = dsS1Sensors['concat'].str.contains('/|\|', regex = True) # Find characters `/` & `\`

dsS1Sensors.loc[dsS1Sensors['specialChar'] == True]                           # Filter for true

dsS1Sensors['concat'].replace('office/study_drawer','study_drawer',inplace=True)
dsS1Sensors['room'].replace('Office/study','Study',inplace=True)
dsS1Sensors['concat'].replace('office/study_lightswitch','study_lightwitch',inplace=True)

dsS1Sensors[0:25]

dsS1Sensors[26:50]

dsS1Sensors[50:76]

Comment: Later the ## values for subActNum will become subActNum_##

dsS1Sensors['subActNumConcat'] = 'subActNum_' + dsS1Sensors['subActNum'].astype(str)

dsS1Sensors[0:10]

subActNumKeyWithStringDict = pd.Series(dsS1Sensors.concat.values, 
                                       dsS1Sensors.subActNumConcat.values).to_dict()
print("subActNum_100: ", subActNumKeyWithStringDict['subActNum_100'])    # Checking the dictionary
print("subActNum_101: ", subActNumKeyWithStringDict['subActNum_101']) 

subActNumKeyWithStringJson = json.dumps(subActNumKeyWithStringDict)
f = open("subActNumKeyWithStringDict.json","w")
f.write(subActNumKeyWithStringJson)
f.close()

subActNumKeyWithEnergyDict = pd.Series(dsS1Sensors.reqElectricity.values, 
                                    dsS1Sensors.subActNumConcat.values).to_dict()
print("subActNum_100: ", subActNumKeyWithEnergyDict['subActNum_100'])    # Checking the dictionary
print("subActNum_101: ", subActNumKeyWithEnergyDict['subActNum_101']) 

subActNumKeyWithEnergyJson = json.dumps(subActNumKeyWithEnergyDict)
f = open("subActNumKeyWithEnergyDict.json","w")
f.write(subActNumKeyWithEnergyJson)
f.close()

# Sub act string key with energy

subActKeyWithEnergyDict = pd.Series(dsS1Sensors.reqElectricity.values, 
                                    dsS1Sensors.concat.values).to_dict()
print("bathroom_door:       ", subActKeyWithEnergyDict['bathroom_door'])    # Checking the dictionary
print("bathroom_exhaustfan: ", subActKeyWithEnergyDict['bathroom_exhaustfan']) 

subActKeyWithEnergyJson = json.dumps(subActKeyWithEnergyDict)
f = open("subActKeyWithEnergyDict.json","w")
f.write(subActKeyWithEnergyJson)
f.close()

seen = {}
dupes = []

for x in dsS1Sensors.concat:
    if x not in seen:
        seen[x] = 1
    else:
        if seen[x] == 1:
            dupes.append(x)
        seen[x] += 1

for i in seen:                  # Adding value counts for dupes
    if (seen[i] > 1 and seen[i] <= 9):
        print(seen[i], " ", i)
    if (seen[i] >= 10):
        print(seen[i], "", i)

dsS1Sensors.drop(columns=['specialChar'], inplace=True)

dsS1Sensors.to_csv('S1Sensors_preprocessed.csv',index = False)

---
