==> 03. S1a SubActivities Preprocessing.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime

---

dsS1 = pd.read_csv('S1activities_data.csv', sep = 'delimiter', header = None, engine = 'python')

a = np.array(dsS1)
a = a.flatten()                                        # Flattening the 2D array to a 1D array
from more_itertools import chunked                     # Importing the required packages
a = list(chunked(a, 5))                                # Chunking the array by a unit of 5 

dimLength = []
i = 0
while i < len(a):
    dimLength.append(len(a[i]))
    i = i + 1

len(dimLength)

### Extracting the Sub-Activities

tempStringList = []                       # Temporary string list
tempStripList = []                        # Temporary 
activityList = []
dateList = []
tempStartTimeList = []
tempEndTimeList = []

i = 0
while i < len(dimLength):
    tempStringList.append(a[i][0])
    tempStripList.append([x.strip() for x in tempStringList[i].split(',')])
    activityList.append(tempStripList[i][0])
    dateList.append(tempStripList[i][1])
    tempStartTimeList.append(tempStripList[i][2])
    tempEndTimeList.append(tempStripList[i][3])
    i = i + 1

numericConcat = []
numericSep = []
actConcat = []
actSep = []
startConcat = []
startSep = []
endConcat = []
endSep = []

i = 0
while i < len(a):
    numericConcat.append(a[i][1])
    numericSep.append([x.strip() for x in numericConcat[i].split(',')])
    actConcat.append(a[i][2])
    actSep.append([x.strip() for x in actConcat[i].split(',')])
    startConcat.append(a[i][3])
    startSep.append([x.strip() for x in startConcat[i].split(',')])
    endConcat.append(a[i][4])
    endSep.append([x.strip() for x in endConcat[i].split(',')])
    i = i + 1

dateStringList = []

i = 0
while i < len(a):
    for x in range(len(numericSep[i])):
        dateStringList.append(dateList[i])
    i = i + 1

len(dateStringList)

flat_numericSep = []                                    # Constructing an array 
for sublist in numericSep:
    for item in sublist:
        flat_numericSep.append(item)

flat_actSep = []
for sublist in actSep:
    for item in sublist:
        flat_actSep.append(item)

flat_startSep = []
for sublist in startSep:
    for item in sublist:
        flat_startSep.append(item)

flat_endSep = []
for sublist in endSep:
    for item in sublist:
        flat_endSep.append(item)

len(flat_numericSep), len(flat_actSep), len(flat_startSep), len(flat_endSep)

zippedList = list(zip(flat_numericSep, flat_actSep, dateStringList, flat_startSep, flat_endSep))
ds = pd.DataFrame(zippedList, columns = ['subActNum', 'subAct', 'date', 'startTime', 'endTime'])

start = (ds.date + " " + ds.startTime)
end = (ds.date + " " + ds.endTime)

startList = []
endList = []

i = 0
while i < len(start):
    startList.append(datetime.strptime(start[i], '%m/%d/%Y %H:%M:%S'))
    endList.append(datetime.strptime(end[i], '%m/%d/%Y %H:%M:%S'))
    i = i + 1

zippedList = list(zip(flat_numericSep, flat_actSep, startList, endList))
ds = pd.DataFrame(zippedList, columns = ['subActNum', 'subAct', 'start', 'end'])
ds.head(n=5)

ds.head(n=4)

ds['subAct'].replace('Office/study', 'Study', inplace=True)

ds.info()

ds['subActNum'] = pd.to_numeric(ds.subActNum)

## Considerations for 'duplicate' sub activities
The following duplicate sensors were identified in section 1

**Create table**
<br>

3   kitchen_lightswitch <br>
4   kitchen_burner <br>
2   livingroom_lightswitch <br>
7   kitchen_drawer <br>
3   kitchen_refrigerator <br>
15  kitchen_cabinet <br>
2   kitchen_door <br>
5   bedroom_drawer <br>
2   bathroom_medicinecabinet <br>
2   bathroom_cabinet <br>


| Sensor  Count    | Descriptor               |
| :-----------:    | ----------               |
|      3           | kitchen_lightswitch      |
|      4           | kitchen_burner           |
|      2           | livingroom_lightswitch   |
|      7           | kitchen_drawer           |
|      3           | kitchen_refrigerator     |
|      15          | kitchen_cabinet          |
|      2           | kitchen_door             |
|      5           | bedroom_drawer           |
|      2           | bathroom_medicinecabinet |
|      2           | bathroom_cabinet         |

# Add explanation as to how these will be dealt with

* From `S1Sensors_preprocessed.csv` I need to find the numbers associated with these - SUBACTNUM
* 'Arbitrarily' choose one number to represent all of them
* Fill them all with that one number

**For dicussion later** - The kitchen cabinets have 15 sensors, say one of these cabinets just contains a blender, so there will always be a one-to-one between the cabinet opening and the blendor being used. This isn't particularly helpful to us. Also, more sensors could always be added... Better to do analysis with less (explain?)



dsS1Sensors = pd.read_csv('S1Sensors_preprocessed.csv', engine = 'python')     
# Note for discussion, S1Sensors_preprocessed itself should no longer be modified 
# EXAMPLE, going back upstream to remove the 'dupes' would perhaps cause an lack of traceability downstream later
# Importing S1Sensors_preprecessed now has been done to inform how to remove 'dupes' from the current ds

dsS1Sensors.head(n=1)

dsS1Sensors.info()

ds.head(n=1)

ds.info()

Replace the `ds.subAct` values with `dsS1Sensors.concat` values using the `subActNum` key
* Join OR
* Dict? USE THIS - cleaner

subActNumKeyWithStringDict = pd.Series(dsS1Sensors.concat.values, 
                                       dsS1Sensors.subActNum.values).to_dict()

print("100: ", subActNumKeyWithStringDict[100])    # Checking the dictionary
print("101: ", subActNumKeyWithStringDict[101])

subActNumKeyWithStringDictNoDupes = {}

for key,value in subActNumKeyWithStringDict.items():
    if value not in subActNumKeyWithStringDictNoDupes.values():
        subActNumKeyWithStringDictNoDupes[key] = value

#test = ds.replace({"subAct": subActNumKeyWithStringDict})
#ds['check'] = ds['subActNum'].replace(subActNumKeyWithStringDict.values())
ds['subAct'] = ds['subActNum'].map(subActNumKeyWithStringDict)

ds.head(n=5)

seen = {}
dupes = []

for x in ds.subActNum:
    if x not in seen:
        seen[x] = 1
    else:
        if seen[x] == 1:
            dupes.append(x)
        seen[x] += 1

#for i in seen:                  # Adding value counts for dupes
#    if (seen[i] > 1 and seen[i] <= 9):
#        print(seen[i], "  ", i)
#    if (seen[i] >= 10 and seen[i] <= 99):
#        print(seen[i], " ", i)
#    if (seen[i] >= 100):
#        print(seen[i], "", i)

ds[ds['subAct'].str.contains("kitchen_refrigerator")]  #ADD UNIQUE

subActStringWithNumDictNoDupes = dict([(value, key) for key, value in subActNumKeyWithStringDictNoDupes.items()]) 
# PRINT example entries subActStringWithNumDictNoDupes

ds['subActNum'] = ds['subAct'].map(subActStringWithNumDictNoDupes)

ds.head(n=5)

ds[ds['subAct'].str.contains("kitchen_refrigerator")]  #ADD UNIQUE

subActNumKeyWithStringDictNoDupes

#test = ds.replace({"subAct": subActNumKeyWithStringDict})
#ds['check'] = ds['subActNum'].replace(subActNumKeyWithStringDict.values())
#ds['subAct'] = ds['subActNum'].map(subActNumKeyWithStringDict)

ds.to_csv('S1SubActivities_preprocessed.csv', index = False)

---
