==> 04. S1a SubActivities Added Time Range.ipynb <==
import pandas as pd
import numpy as np
import time
from datetime import datetime

ds = pd.read_csv('S1SubActivities_preprocessed.csv', index_col = None) 

ds.head(n=10)

ds['subAct'].replace('Office/study', 'Study', inplace=True) #req??????

ds.dtypes

# get min/max
initialTimeStart = min(ds.start)
finalTimeStart = max(ds.start)
print("Initial Time for START:", initialTimeStart, '\n' "Final Time for START:", finalTimeStart)

# get min/max
initialTimeEnd = min(ds.end)
finalTimeEnd = max(ds.end)
print("Initial Time for END:", initialTimeStart, '\n' "Final Time for END:", finalTimeStart)

print(min(ds.start))                                   # Checking the start time minima

print(max(ds.start))                                   # Checking the start time maxima

start = ds.start
end = ds.end

tempTimeStamp = []
i = 0
while i < len(ds):
    tempTimeStamp.append(pd.date_range(start = start[i], end = end[i], freq = 's'))
    i += 1

tempNumTimeStamps = []
i = 0
while i < len(ds):
    tempNumTimeStamps.append(len(tempTimeStamp[i]))
    i += 1

ds['actDuration'] = tempNumTimeStamps
ds['timeStampList'] = tempTimeStamp

timeStampVector = ds['timeStampList']

timeStampVector[1]

DTArrayList = []
i = 0
while i < len(ds):
    DTArrayList.append(pd.DatetimeIndex.to_pydatetime(timeStampVector[i]))
    i = i + 1

ds['timeStampArrayList'] = DTArrayList

ds.head(n=4)

ds.to_csv('S1SubActivities_timeStampRanges.csv',index=False)
