==> 05. S1a SubAct Duration Temporal Features.ipynb <==
import pandas as pd
import numpy as np
import time
from datetime import datetime
import altair as alt
import pandas as pd

#https://vega.github.io/vega-lite/tutorials/explore.html
# USE THIS COLOUR PATETTE (end of page)
# https://altair-viz.github.io/gallery/selection_histogram.html
# https://altair-viz.github.io/user_guide/interactions.html
# https://altair-viz.github.io/gallery/iowa_electricity.html

ds = pd.read_csv('S1SubActivities_preprocessed.csv', index_col = None) 
ds.start = pd.to_datetime(ds.start, format='%Y-%m-%d %H:%M:%S')
ds.end = pd.to_datetime(ds.end, format='%Y-%m-%d %H:%M:%S')
ds = ds.sort_values('start')
ds.reset_index(drop = True, inplace = True)

ds.head(n=2)

%run add_DAY_WDWE_phaseI.py

ds = add_DAY_WDWE_phaseI(ds)

ds.head(n=2)

ds.info()

1. Set specified column as index
2. Extract dayofweek from index (ds.index.dayofweek.astype(str)), as this as an attribute
3. Replace '0' with 'mon' and so on
4. Duplic

# Unique subActNum list
unique_list = []
for x in ds.subAct:                     # Traverse for all elements 
    if x not in unique_list:            # Check if exists in unique_list or not 
        unique_list.append(x)     

for subAct in unique_list:
    
    fig = px.histogram(ds[ds['subAct']==subAct], x="HOUR", y="HOUR", color="WDWE",
                   marginal="box", # or violin, rug
                   hover_data=dsCheck.columns)
    fig.show()

import plotly.express as px
tips = px.data.tips()
fig = px.histogram(dsCheck, x="HOUR", y="HOUR", color="WDWE",
                   marginal="box", # or violin, rug
                   hover_data=dsCheck.columns)
fig.show()

dsCheck = ds[ds['subAct']=='bathroom_cabinet']

dsCheck.head(n=5)

len(dsCheck)

import plotly.graph_objects as go

import numpy as np

x = dsCheck['deltaStartEnd']

fig = go.Figure(data=[go.Histogram(x=x)])
fig.show()

import plotly.graph_objects as go

import numpy as np

x = dsCheck['HOUR']

fig = go.Figure(data=[go.Histogram(x=x)])
fig.show()

import plotly.graph_objects as go

import numpy as np

x = dsCheck['HOUR']

fig = go.Figure(data=[go.Histogram(x=x, histnorm='probability')])
fig.show()

alt.Chart(source).transform_timeunit(Month='month(date)').transform_joinaggregate(
    mean_temp='mean(temp_max)', 
    groupby=['Month'].transform_bin(['bin_max', 'bin_min'], 'temp_max').transform_aggregate(
    value='count()', groupby=['Month', 'mean_temp', 'bin_min', 'bin_max']).transform_impute(
    impute='value', groupby=['Month', 'mean_temp'], key='bin_min', value=0).mark_area(
    interpolate='monotone', fillOpacity=0.8, stroke='lightgray', strokeWidth=0.5).encode(
    alt.X('bin_min:Q', 
          bin='binned', 
          title='Maximum Daily Temperature (C)'),
    alt.Y('value:Q', 
          scale=alt.Scale(range = [step, -step * overlap]), axis=None), 
    alt.Fill('mean_temp:Q',
        legend=None,
        scale=alt.Scale(domain=[30, 5], scheme='redyellowblue')),
    alt.Row('Month:T',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B'))).properties(
    bounds='flush', title='Seattle Weather', height=step).configure_facet(
    spacing=0).configure_view(
    stroke=None).configure_title(
    anchor='end')

step = 40
overlap = 0

alt.Chart(ds
).transform_bin(
    ['bin_max', 'bin_min'], 'deltaStartEnd'
).transform_aggregate(
    value='count()', groupby=['HOUR', 'deltaStartEnd', 'bin_min', 'bin_max']
).transform_impute(
    impute='value', groupby=['HOUR', 'deltaStartEnd'], key='bin_min', value=0
).mark_area(
    interpolate='monotone',
    fillOpacity=0.8,
    stroke='black',
    strokeWidth=0.5
).encode(
    alt.X('bin_min:Q', bin='binned', title='Delta'),
    alt.Y(
        'value:Q',
        scale=alt.Scale(range=[step, -step * overlap]),
        axis=None
    ),
    alt.Row(
        'HOUR:N',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B')
    )
).properties(
    bounds='flush', title='67-100', height=step
).configure_facet(
    spacing=0
).configure_view(
    stroke=None
).configure_title(
    anchor='end'
)


step = 40
overlap = 0

alt.Chart(ds
).transform_joinaggregate(
    mean_delta='mean(deltaStartEnd)', groupby=['subAct']
).transform_bin(
    ['bin_max', 'bin_min'], 'deltaStartEnd'
).transform_aggregate(
    value='count()', groupby=['subAct', 'deltaStartEnd', 'bin_min', 'bin_max']
).transform_impute(
    impute='value', groupby=['subAct', 'deltaStartEnd'], key='bin_min', value=0
).mark_area(
    interpolate='monotone', fillOpacity=0.8, stroke='black', strokeWidth=0.5
).encode(
    alt.X('bin_min:Q', bin='binned', title='Delta'),
    alt.Y(
        'value:N',
        scale=alt.Scale(range=[step, -step * overlap]),
        axis=None),
    alt.Fill('mean_delta:Q',
        legend=None,
        scale=alt.Scale(domain=[30000, 5], scheme='redyellowblue')),
    alt.Row(
        'HOUR:N',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B')
    )
).properties(
    bounds='flush', title='67-100', height=step
).configure_facet(
    spacing=0
).configure_view(
    stroke=None
).configure_title(
    anchor='end'
).interactive()

step = 40
overlap = 0

alt.Chart(ds
).transform_joinaggregate(
    mean_delta='mean(deltaStartEnd)', groupby=['subAct']
).transform_bin(
    ['bin_max', 'bin_min'], 'deltaStartEnd'
).transform_aggregate(
    value='count()', groupby=['HOUR', 'deltaStartEnd', 'bin_min', 'bin_max']
).transform_impute(
    impute='value', groupby=['subAct', 'deltaStartEnd'], key='bin_min', value=0
).mark_area(
    interpolate='monotone', fillOpacity=0.8, stroke='black', strokeWidth=0.5
).encode(
    alt.X('bin_min:Q', bin='binned', title='Delta'),
    alt.Y(
        'value:N',
        scale=alt.Scale(range=[step, -step * overlap]),
        axis=None),
    alt.Fill('mean_delta:Q',
        legend=None,
        scale=alt.Scale(domain=[30000, 5], scheme='redyellowblue')),
    alt.Row(
        'HOUR:N',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B')
    )
).properties(
    bounds='flush', title='67-100', height=step
).configure_facet(
    spacing=0
).configure_view(
    stroke=None
).configure_title(
    anchor='end'
).interactive()


step = 40
overlap = 0

alt.Chart(ds
).transform_bin(
    ['bin_max', 'bin_min'], 'deltaStartEnd'
).transform_aggregate(
    value='count()', groupby=['HOUR', 'deltaStartEnd', 'bin_min', 'bin_max']
).transform_impute(
    impute='value', groupby=['subAct', 'deltaStartEnd'], key='bin_min', value=0
).mark_area(
    interpolate='monotone',
    fillOpacity=0.8,
    stroke='black',
    strokeWidth=0.5
).encode(
    alt.X('bin_min:Q', bin='binned', title='Delta'),
    alt.Y(
        'value:N',
        scale=alt.Scale(range=[step, -step * overlap]),
        axis=None
    ),
    alt.Row(
        'HOUR:N',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B')
    )
).properties(
    bounds='flush', title='67-100', height=step
).configure_facet(
    spacing=0
).configure_view(
    stroke=None
).configure_title(
    anchor='end'
)


* Title with SubActNum and SubAct
* Start Date as Thu 03 March
* Weekends orange
* Weekdays blue
* 24till2
* 3till5
* 6till8
* 9till11
* 12till14
* 15till17
* 18till20
* 21till23

# Unique subActNum list
unique_list = []
for x in ds.subActNum:                     # Traverse for all elements 
    if x not in unique_list:            # Check if exists in unique_list or not 
        unique_list.append(x)     


step = 20
overlap = 1

alt.Chart(ds).transform_joinaggregate(
    mean_duration = 'mean(deltaStartEnd)', 
    groupby=['subActNum']).transform_bin(['bin_max', 'bin_min'], 'max(deltaStartEnd)').transform_aggregate(
    value='count()', groupby=['subActNum', 'mean_duration', 'bin_min', 'bin_max']).transform_impute(
    impute='value', groupby=['subActNum', 'mean_duration'], key='bin_min', value=0).mark_area(
    interpolate='monotone', fillOpacity=0.8, stroke='lightgray', strokeWidth=0.5).encode(
    alt.X('bin_min:Q', 
          bin='binned', 
          title='Maximum Daily Temperature (C)'),
    alt.Y('value:Q', 
          scale=alt.Scale(range = [step, -step * overlap]), axis=None), 
    alt.Fill('mean_duration:Q',
        legend=None,
        scale=alt.Scale(domain=[30, 5], scheme='redyellowblue')),
    alt.Row('date(start):T',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B'))).properties(
    bounds='flush', title='Seattle Weather', height=step).configure_facet(
    spacing=0).configure_view(
    stroke=None).configure_title(
    anchor='end')


# https://vega.github.io/vega-lite/data/seattle-weather.csv
# 
alt.Chart(source).transform_timeunit(Month='month(date)').transform_joinaggregate(
    mean_temp='mean(temp_max)', 
    groupby=['Month'].transform_bin(['bin_max', 'bin_min'], 'temp_max').transform_aggregate(
    value='count()', groupby=['Month', 'mean_temp', 'bin_min', 'bin_max']).transform_impute(
    impute='value', groupby=['Month', 'mean_temp'], key='bin_min', value=0).mark_area(
    interpolate='monotone', fillOpacity=0.8, stroke='lightgray', strokeWidth=0.5).encode(
    alt.X('bin_min:Q', 
          bin='binned', 
          title='Maximum Daily Temperature (C)'),
    alt.Y('value:Q', 
          scale=alt.Scale(range = [step, -step * overlap]), axis=None), 
    alt.Fill('mean_temp:Q',
        legend=None,
        scale=alt.Scale(domain=[30, 5], scheme='redyellowblue')),
    alt.Row('Month:T',
        title=None,
        header=alt.Header(labelAngle=0, labelAlign='right', format='%B'))).properties(
    bounds='flush', title='Seattle Weather', height=step).configure_facet(
    spacing=0).configure_view(
    stroke=None).configure_title(
    anchor='end')

ds.head(n=5)

# https://altair-viz.github.io/gallery/binned_heatmap.html
alt.Chart(ds).mark_rect().encode(
    alt.X('DAY'),
    alt.Y('subAct'),
    alt.Color('deltaStartEnd:Q', scale=alt.Scale(scheme='greenblue'))
).properties(width=550).interactive()

# DS Norm
ds['durationNorm'] = (ds['deltaStartEnd'] - min(ds['deltaStartEnd'])
                     )/(max(ds['deltaStartEnd']) - min(ds['deltaStartEnd']))

ds.head(n=5)

# https://altair-viz.github.io/gallery/binned_heatmap.html
alt.Chart(ds).mark_rect().encode(
    alt.X('DAY'),
    alt.Y('subAct'),
    alt.Color('durationNorm:Q', scale=alt.Scale(scheme='inferno'))
).properties(width=550).interactive()

alt.Chart(ds).mark_rect().encode(
    alt.X('date(start):O'),
    alt.Y('subAct'),
    alt.Color('durationNorm:Q', scale=alt.Scale(scheme='inferno'))
).properties(width=550).interactive()

alt.Chart(ds).mark_rect().encode(
    alt.X('date(start):O'),
    alt.Y('subAct'),
    alt.Color('mean(deltaStartEnd)', scale=alt.Scale(scheme='inferno'))
).properties(width=550).interactive()


alt.Chart(ds).mark_rect().encode(
    alt.X('hours(start):O'),
    alt.Y('subAct'),
    alt.Color('durationNorm:Q', scale=alt.Scale(scheme='inferno')),
    row = 'DAY'
).properties(width=550).interactive()

alt.Chart(ds).mark_boxplot().encode(
    x = 'durationNorm',
    y = 'subAct',
).properties(width = 550).interactive()

alt.Chart(ds).mark_bar().encode(
    alt.X('durationNorm', bin = alt.Bin(maxbins=100)),
    y = 'count(subAct)',
    row = 'subAct'
).properties(width = 550).interactive()

#https://altair-viz.github.io/gallery/stacked_bar_chart_sorted_segments.html
alt.Chart(ds).mark_bar().encode(
    x = 'count(subAct)',
    y = 'subAct',
    color = 'WDWE',
    column = 'WDWE' 
)

#https://altair-viz.github.io/gallery/stacked_bar_chart_sorted_segments.html
alt.Chart(ds).mark_bar().encode(
    x = 'count(subAct)',
    y = 'subAct',
    color = 'DAY',
    row = 'WDWE' 
)

#https://altair-viz.github.io/gallery/stacked_bar_chart_sorted_segments.html
alt.Chart(ds).mark_bar().encode(
    x = 'count(subAct)',
    y = 'subAct',
    color = 'DAY',
    column = 'WDWE' 
)

#https://altair-viz.github.io/gallery/stacked_bar_chart_sorted_segments.html
# https://altair-viz.github.io/gallery/multiple_marks.html
alt.Chart(ds).mark_line(point=True).encode(
    y = 'count(subAct)',
    x = 'subAct',
    color = 'DAY',
    row = 'WDWE' 
)
# PLOT AS SOLID FILL

# https://altair-viz.github.io/gallery/layered_area_chart.html
alt.Chart(ds).mark_area(opacity = 0.8).encode(
    y = 'count(subAct)',
    x = alt.X('subAct', axis = alt.Axis(labelAngle = -45)),
    color = 'DAY',
    row = 'WDWE' 
)
# PLOT AS SOLID FILL

#https://altair-viz.github.io/gallery/stacked_bar_chart_sorted_segments.html
input_dropdown = alt.binding_select(options=['Mon','Tue','Wed','Thu','Fri','Sat','Sun'])
selection = alt.selection_single(fields=['DAY'], bind=input_dropdown, name='Test Name')
color = alt.condition(selection,
                    alt.Color('WDWE:Q', legend=None),
                    alt.value('lightgray'))

alt.Chart(ds).mark_bar().encode(
    x = 'count(subAct)',
    y = 'subAct',
    color = 'DAY',
    column = 'WDWE' 
).add_selection(
    selection
).transform_filter(
    selection
)


# https://altair-viz.github.io/gallery/natural_disasters.html
alt.Chart(ds).mark_circle(
    opacity=0.8,
    stroke='black',
    strokeWidth=1
).encode(
    alt.X('date(start):O', axis=alt.Axis(labelAngle=0)),
    alt.Y('subAct:N', axis=alt.Axis(labelAngle=0)),
    alt.Size('count(subAct):Q',
        scale=alt.Scale(range=[0, 2000]),
        legend=alt.Legend(title='Annual Global Deaths')
    ),
    alt.Color('WDWE:N', legend=None)
).properties(
    width=600,
    height=800
).interactive()

#date(start):O
    
    # https://altair-viz.github.io/gallery/natural_disasters.html
alt.Chart(ds).mark_circle(
    opacity=0.8,
    stroke='black',
    strokeWidth=1
).encode(
    alt.X('date(start):T', axis=alt.Axis(labelAngle=0)),
    alt.Y('subAct:N', axis=alt.Axis(labelAngle=0)),
    alt.Size('count(subAct):Q',
        scale=alt.Scale(range=[0, 2000]),
        legend=alt.Legend(title='Annual Global Deaths')
    ),
    alt.Color('WDWE:N', legend=None)
).properties(
    width=600,
    height=800
).interactive()

#date(start):O
    
    # https://altair-viz.github.io/gallery/natural_disasters.html
alt.Chart(ds).mark_circle(
    opacity=0.8,
    stroke='black',
    strokeWidth=1
).encode(
    alt.X('start:T', axis=alt.Axis(labelAngle=0)),
    alt.Y('subAct:N', axis=alt.Axis(labelAngle=0)),
    alt.Size('count(subAct):Q',
        scale=alt.Scale(range=[0, 1000]),
        legend=alt.Legend(title='Annual Global Deaths')
    ),
    alt.Color('WDWE:N', legend=None)
).properties(
    width=600,
    height=800
).interactive()

#https://altair-viz.github.io/gallery/stripplot.html
stripplot =  alt.Chart(ds, width = 40).mark_circle(size=8).encode(
    x=alt.X(
        'jitter:N',
        title=None,
        axis=alt.Axis(values=[0], ticks=True, grid=False, labels=False),
        scale=alt.Scale(),
    ),
    y=alt.Y('deltaStartEnd'),
    color=alt.Color('DAY:N', legend=None),
    column=alt.Column(
        'subAct:N',
        header=alt.Header(
            labelAngle=-90,
            titleOrient='top',
            labelOrient='bottom',
            labelAlign='right',
            labelPadding=0,
        ),
    ),
).transform_calculate(
    # Generate Gaussian jitter with a Box-Muller transform
    jitter='sqrt(-2*log(random()))*cos(2*PI*random())'
).configure_facet(
    spacing=0
).configure_view(
    stroke=None
)

stripplot

alt.Chart(
    ds,
    title="2010 Daily High Temperature (F) in Seattle, WA"
).mark_rect().encode(
    x='hoursminutes(start)',
    y='subAct:O',
    color=alt.Color('durationNorm', scale=alt.Scale(scheme="inferno")),
    #tooltip=[
     #   alt.Tooltip('monthdate(date):T', title='Date'),
      #  alt.Tooltip('max(temp):Q', title='Max Temp')
    #]
).properties(width=550)
