==> 07. BRANCH S1a SubActivities Collapse into Minutes.ipynb <==
## Collapse into minutes
* Use in-built Pandas function
* `resample('min').max()` 

import pandas as pd
import numpy as np
from datetime import datetime

ds = pd.read_csv('S1SubActivities_timeRangeBoolean.csv', index_col = 'duration')
ds.index = pd.to_datetime(ds.index)

ds.head(n=2)

ds.tail(n=2)

ds = ds.resample('min').max()

ds.head(n=2)

ds.tail(n=2)

len(ds)

ds.to_csv('S1SubActivities_timeRangeBooleanMinutes.csv', index = 'duration')

ds = ds.dropna(how='all')

len(ds)

ds.to_csv('S1SubActivities_timeRangeBooleanMinutesDropNA.csv', index = 'duration')

---
