==> 08. BRANCH S1a SubActivities Remove Duplicate Attributes.ipynb <==
import pandas as pd
import numpy as np

---

## Collapsing 'Duplicate' Attributes
* 

ds = pd.read_csv('S1SubActivities_timeRangeBooleanMinutesDropNA.csv', index_col = 'duration')
ds.index = pd.to_datetime(ds.index)

dsS1Sensors = pd.read_csv('S1Sensors_preprocessed.csv', index_col = None)     # Importing the dataset

dsS1Sensors.head(n=5)

seen = {}
dupes = []

for x in dsS1Sensors.concat:
    if x not in seen:
        seen[x] = 1
    else:
        if seen[x] == 1:
            dupes.append(x)
        seen[x] += 1

dupes

dsSensorDrop = dsS1Sensors.drop(columns = ['subActNum', 'room', 'activity', 'reqElectricity'])
dsSensorDrop.head(n=5)

sensorDict = pd.Series(dsSensorDrop.concat.values, dsSensorDrop.subActNumConcat.values).to_dict()
sensorDict["subActNum_100"]

ds.rename(columns = sensorDict, inplace=True)

ds.head(n=2)

colList = []

for i in ds.columns:
    colList.append(i)

len(colList)

ds = ds.groupby(ds.columns, axis=1).sum()

ds.head(n=5)

colList = []

for i in ds.columns:
    colList.append(i)

len(colList)

maxValues = ds.max()                                # Checking for max values in df
maxValues.max()                                     # The maximum value of max values

ds = ds.replace(to_replace =[2, 3, 4, 5], value = 1)

maxValues = ds.max()
maxValues.max()

ds.to_csv('S1SubAct_B_m_NoDupes.csv',index='duration')
