==> 9. S1b SubActivities Plots OLD.ipynb <==
import pandas as pd
import numpy as np
import time
from datetime import datetime
import altair as alt
import pandas as pd
import json
import itertools as it

ds = pd.read_csv('S1SubActivities_preprocessed.csv', index_col = None)
ds.start = pd.to_datetime(ds.start, format='%Y-%m-%d %H:%M:%S')
ds.end = pd.to_datetime(ds.end, format='%Y-%m-%d %H:%M:%S')
ds = ds.sort_values('start')
ds.reset_index(drop = True, inplace = True)

ds = ds[:100]
ds.head(30)

test = pd.DataFrame()
test.
test['columnNsme'] = 10
test

ds.head(n=15)

Problem Statement is:
If I switch on the bathroom light switch between 6am and 7am on a weekday, the probability that I will use my razor is X. 

If start EVENTA between X and Y on a WD/WE the PR that I will start EVENTB within Z minutes is Q.

import datetime as dt

testEnd = dt.datetime(2003,3,27,7)
testStart = dt.datetime(2003,3,27,6)
testTimeFrame = dt.timedelta(hours = 1)

def eventPR(ds, eventAStart, periodStart, periodEnd, dayType, eventB, timeFrame):
    tempDF = ds[(ds.start < periodEnd + timeFrame) & (ds.start > periodStart)]
    return tempDF
    
eventPR(ds, 0, testStart, testEnd, 0, 0, testTimeFrame)
                                                    

Above is A

* Title with SubActNum and SubAct
* Start Date as Thu 03 March
* Weekends orange
* Weekdays blue
* 24till2
* 3till5
* 6till8
* 9till11
* 12till14
* 15till17
* 18till20
* 21till23

# Unique subActNum list
unique_list = []
for x in ds.subActNum:                     # Traverse for all elements 
    if x not in unique_list:            # Check if exists in unique_list or not 
        unique_list.append(x)     

---

iteration_2n = it.combinations(ds.subActNum, 2)
array_2n = np.array(list(iteration_2n))
len(array_2n)

array_2n[0:4]

dsCombin2n = pd.DataFrame({'EventA': array_2n[:, 0], 
                           'EventB': array_2n[:, 1]})

dsCombin2n.head(n=5)

Above is B

ds.head(n=5)

WHERE event A is followed by event B

IF (event A end < event B start)

SUM (event B start - event A end) = delta

Add delta attribute to df

RETURN df

import datetime as dt

def nearest_neighbours(events, n=1, delta_threshold=dt.timedelta(-99)):
    nns = []
    for row in events.itertuples():
        #print(row)
        start_time = getattr(row, 'start')
        end_time = getattr(row, 'end')
        subActNum = getattr(row, 'subActNum')
        #print(subActNum)
        row_index = getattr(row, 'Index')
        nn = events[(events.start >= start_time) & (events.index != row_index) & (end_time - events.start > delta_threshold)][:n]
        ordered = pd.DataFrame()
        ordered['EventA'] = subActNum
        #print(ordered['EventA'])
        ordered['EventB'] = nn['subActNum']
        ordered['EvA_Start'] = start_time
        ordered['EvA_End'] = end_time
        ordered['EvB_Start'] = nn['start']
        ordered['EvB_End'] = nn['end']
        nns.append(ordered)
  
    #print(nns)
    return pd.concat(nns)

def id_delta(events):
    df = events.copy()
    return df

#id_delta(ds).head()
nearest_neighbours(ds, 5, dt.timedelta(0,-10))

ds
