==> 11. S1b Combinational Evaluation.ipynb <==
import pandas as pd
import numpy as np
import json
import itertools as it

---

## The Sensor Data
### Importing Sensor Data
The file `S1sensors.csv` contains a tabulated values for sensor ID (numeric), room and sensor type.

dsS1Sensors = pd.read_csv('S1Sensors_preprocessed.csv', index_col = False)

dsS1Sensors.head(n=5)                                         # Checking the dataset

---

# Relationships

### Intersection
<pre>
|----EventA----|  <br>
            |---EventB---|   
</pre>

### Separate
<pre>
|----EventA----|  <br>
                    |---EventB---|   
</pre>

### Enclosed
<pre>
|----------EventA----------|  <br>
           |---EventB---|   
</pre>

### Equal
<pre>
       |---EventA---|  <br>
       |---EventB---|   
</pre>

### More?

A & B & C

A &rightarrow; B


---

iteration_2n = it.combinations(dsS1Sensors.subActNum, 2)
array_2n = np.array(list(iteration_2n))
len(array_2n)

array_2n[0:4]

dsCombin2n = pd.DataFrame({'EventA': array_2n[:, 0], 
                           'EventB': array_2n[:, 1]})

dsCombin2n.head(n=5)

dsCombin2n.tail(n=5)

---

iteration_3n = it.combinations(dsS1Sensors.subActNum, 3)
array_3n = np.array(list(iteration_3n))
len(array_3n)

array_3n[0:4]

dsCombin3n = pd.DataFrame({'EventA': array_3n[:, 0], 
                           'EventB': array_3n[:, 1],
                           'EventC': array_3n[:, 2]})

dsCombin3n.head(n=5)

# Check all durations for outliers

| EventA| EventB| Ev.A Start| Ev.A End| Ev.B Start|Ev.B End|Delta   | Descriptor | WendWday|Hour  |TimeofDay|
|:-----:|:-----:|:--------: | :------:|:--------: |:------:|:------:| :------:   | :------:|:----:|:------: |
| 100   | 101   | 10am      | dateTime| dateTime  |dateTime|-4      | Gap        | weekday | 6:00 |6Till8   |
| 100   | 101   | 11am      | dateTime| dateTime  |dateTime|92      | Union      | weekend |12:00 |12Till14 |
